requirements.txt
matplotlib
numpy
networkx
scikit-learn
torch
fair-esm
pandas
rdkit
tqdm
torch_geometric